fix: rename commom to common
