+++ app/src/hooks/courses/useListCourses.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { useJsonFromResponse } from "@/commom";
+import { useJsonFromResponse } from "@/common";
 import NestAPI from "@/api";
 import { CourseControllerFindRequest } from "@/sdk";
