+++ app/src/hooks/faculty/useListFaculties.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { useJsonFromResponse } from "@/commom";
+import { useJsonFromResponse } from "@/common";
 import NestAPI from "@/api";
 
 export const useListFaculties = () => {
